Refactor bank card number generation and validation logic
This commit includes a series of related changes aiming at improving the code organization and readability of the bank card-related logic in the jmp-cloud-bank-impl module. The primary changes are:

- The CardNumberSupplier and CardNumberValidator classes were renamed to CardNumberSupplier and LuhnValidator, respectively, to better reflect their roles.

- Their test classes were accordingly renamed to CardNumberSupplierSpec and LuhnValidatorSpec.

- The path of the CardNumberSupplier and LuhnValidator classes were changed to reside in the jmp.cloud.bank.cardnumber package for better organization.

- BankCloudImpl class was renamed to BankImpl to reflect that these changes are not necessarily cloud-specific. Its test class, BankCloudImplSpec also got renamed to BankImplSpec.

- Updates on pom.xml files to let maven-compiler-plugin collect information during the build and make the build more efficient.

All these refactorings are aimed to make the code cleaner and easy to navigate. Moreover, these changes have no effect on the functionality of the code.
