+++ jmp-cloud-bank-impl/pom.xml
@@ -30,4 +30,12 @@
 		</dependency>
     </dependencies>
 
+	<build>
+		<plugins>
+			<plugin>
+				<artifactId>maven-compiler-plugin</artifactId>
+			</plugin>
+		</plugins>
+	</build>
+
 </project>
