+++ jmp-cloud-bank-impl/src/main/java/com/epam/engx/jmp/cloud/bank/BankImpl.java
@@ -1,6 +1,7 @@
-package com.epam.engx.jmp.cloud.bank.impl;
+package com.epam.engx.jmp.cloud.bank;
 
 import com.epam.engx.jmp.api.Bank;
+import com.epam.engx.jmp.cloud.bank.cardnumber.CardNumberSupplier;
 import com.epam.engx.jmp.dto.BankCard;
 import com.epam.engx.jmp.dto.BankCardType;
 import com.epam.engx.jmp.dto.CreditBankCard;
@@ -9,14 +10,14 @@ import com.epam.engx.jmp.dto.User;
 
 import java.util.Objects;
 
-public final class BankCloudImpl implements Bank {
+public final class BankImpl implements Bank {
 
 	@Override
 	public BankCard createBankCard(User user, BankCardType type) {
 		Objects.requireNonNull(user, "User cannot be null");
 		Objects.requireNonNull(type, "Bank card type cannot be null");
 
-		var cardNumberSupplier = new LuhnCardNumberSupplier();
+		var cardNumberSupplier = new CardNumberSupplier();
 		var cardNumber = cardNumberSupplier.get();
 
 		return switch (type) {
