--- jmp-cloud-bank-impl/src/main/java/com/epam/engx/jmp/cloud/bank/impl/CardNumberSupplier.java
@@ -1,7 +1,0 @@
-package com.epam.engx.jmp.cloud.bank.impl;
-
-import java.util.function.Supplier;
-
-@FunctionalInterface
-public interface CardNumberSupplier extends Supplier<String> {
-}
