+++ jmp-dto/pom.xml
@@ -21,4 +21,13 @@
 			<artifactId>groovy</artifactId>
 		</dependency>
 	</dependencies>
+
+	<build>
+		<plugins>
+			<plugin>
+				<artifactId>maven-compiler-plugin</artifactId>
+			</plugin>
+		</plugins>
+	</build>
+
 </project>
